// Drink-and-delight-ProductAngular/src/app/model/productstock.ts
export class ProductStock {
    stockId: string;
    orderId: string;
    name: string;
    price_per_unit: number;
    quantityValue: number;
    quantityUnit: number;
    price: number;
    warehouseId: string;
    deliveryDate: Date;
    manuDate: Date;
    expiryDate: Date;
    qualityCheck: string;
    processDate: string;

   }